# api/openapispec/swagger.yaml
definitions:
...
  unstructured.Unstructured:
    properties:
      object:
        additionalProperties: true
        description: |-
          Object is a JSON compatible map with string, float, int, bool, []interface{}, or
          map[string]interface{}
          children.
        type: object
    type: object
info:
  contact: {}
  title: Karpor
  version: "1.0"
paths:
...
  /rest-api/v1/insight/detail:
    get:
      description: This endpoint returns a Kubernetes resource by name, namespace,
        cluster, apiVersion and kind.
      parameters:
      - description: The format of the response. Either in json or yaml. Default to
          json
        in: query
        name: format
        type: string
      - description: The specified cluster name, such as 'example-cluster'
        in: query
        name: cluster
        type: string
      - description: The specified apiVersion, such as 'apps/v1'. Should be percent-encoded
        in: query
        name: apiVersion
        type: string
      - description: The specified kind, such as 'Deployment'
        in: query
        name: kind
        type: string
      - description: The specified namespace, such as 'default'
        in: query
        name: namespace
        type: string
      - description: The specified resource name, such as 'foo'
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Unstructured object
          schema:
            $ref: '#/definitions/unstructured.Unstructured'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetDetail returns a Kubernetes resource by name, namespace, cluster,
        apiVersion and kind.
      tags:
      - insight
  /rest-api/v1/insight/events:
    get:
      description: This endpoint returns events for a Kubernetes resource YAML by
        name, namespace, cluster, apiVersion and kind.
      parameters:
      - description: The specified cluster name, such as 'example-cluster'
        in: query
        name: cluster
        type: string
      - description: The specified apiVersion, such as 'apps/v1'. Should be percent-encoded
        in: query
        name: apiVersion
        type: string
      - description: The specified kind, such as 'Deployment'
        in: query
        name: kind
        type: string
      - description: The specified namespace, such as 'default'
        in: query
        name: namespace
        type: string
      - description: The specified resource name, such as 'foo'
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of events
          schema:
            items:
              $ref: '#/definitions/unstructured.Unstructured'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "405":
          description: Method Not Allowed
          schema:
            type: string
        "429":
          description: Too Many Requests
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: GetEvents returns events for a Kubernetes resource by name, namespace,
        cluster, apiVersion and kind.
      tags:
      - insight
...
swagger: "2.0"